requirements.txt
uvicorn
fastapi
python-dotenv
requests
openai
dart-fss
pymysql
mysql-connector-python
OpenDartReader